make sure tests passes
